move tag to pkg
