+++ tls/handlers/handlers.go
@@ -10,9 +10,9 @@ import (
 	"github.com/jmpsec/osctrl/pkg/environments"
 	"github.com/jmpsec/osctrl/pkg/logging"
 	"github.com/jmpsec/osctrl/pkg/nodes"
+	"github.com/jmpsec/osctrl/pkg/tags"
 	"github.com/jmpsec/osctrl/queries"
 	"github.com/jmpsec/osctrl/settings"
-	"github.com/jmpsec/osctrl/tags"
 	"github.com/jmpsec/osctrl/version"
 )
